rewrite `ef` to replace `funced`, not wrap it
The way `funced` currently works results in data loss: specifically, because of
its reliance on `fish_indent`, `funced` will strip comments & any code outside
the main function definition from any file it touches. I'm reluctant to call
this a bug in fish but it's definitely not behaviour I want, so here's a
minimal reimplementation that behaves the way I prefer.