explicitly define `SHELL` in Makefiles
"There seems to be some issue with using the bash built-in `cd` within
Makefiles. When invoked as `/bin/sh` within Makefiles, bash seems to
have a problem changing directory via the `cd` command. A work-around
for this is to define `SHELL=/bin/bash` within your Makefile. This is
believed to be a bug in bash."