update screenshot of prompt
This prompt hasn't changed much since 2019; I keep writing ones that are
more elaborate and even multi-line and so far nothing really beats this
combination of utility and (lack of) size.

The one relevant feature it's really missing is a timestamp: it's often
useful to check when a command that didn't touch the file system was
issued, or when a long process started. The most trivial implementation
would technically only print the time the previous command finished; to
improve on that would require re-writing each prompt once its command
is issued, and that's too much complexity for me, especially since the
only terminal emulator I use these days (iTerm2) shows timestamps when
you hover over the right margin of the window.

Will update again in 2031, lol